library refactor cross correlograms
