+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { MuiThemeProvider } from '@material-ui/core';
-import RecordingSelectionContext, { defaultRecordingSelection, recordingSelectionReducer } from 'contexts/RecordingSelectionContext';
+import { RecordingSelectionContext, defaultRecordingSelection, recordingSelectionReducer } from 'libraries/RecordingSelectionContext';
 import { getFigureData, useWindowDimensions } from 'figurl';
 import SetupUrlState from 'figurl/SetupUrlState';
 import { defaultUnitSelection, UnitSelectionContext, unitSelectionReducer } from 'libraries/UnitSelectionContext';
