+++ src/libraries/CrossCorrelograms/CrossCorrelogramsViewData.ts
