+++ src/libraries/CrossCorrelograms/index.ts
@@ -1,0 +1,4 @@
+export { default as CrossCorrelogramsView, determinePlotSizeForSquareMatrixGrid } from './CrossCorrelogramsView'
+export { isCrossCorrelogramsViewData } from './CrossCorrelogramsViewData'
+export type { CrossCorrelogramsViewData } from './CrossCorrelogramsViewData'
+
