+++ src/libraries/RecordingSelectionContext/index.ts
@@ -1,0 +1,3 @@
+export { default as RecordingSelectionContext, defaultRecordingSelection, defaultZoomScaleFactor, recordingSelectionReducer, useRecordingSelection, useRecordingSelectionTimeInitialization, useSelectedElectrodes, useTimeFocus, useTimeRange } from './RecordingSelectionContext';
+export type { PanDirection, ZoomDirection } from './RecordingSelectionContext';
+
