+++ src/views/common/Animation/AnimationUtilities/useTimeWindowSyncing.ts
@@ -1,4 +1,4 @@
-import { useTimeRange } from 'contexts/RecordingSelectionContext'
+import { useTimeRange } from 'libraries/RecordingSelectionContext'
 import React, { useEffect } from "react"
 import { AnimationState, AnimationStateAction } from 'views/common/Animation/AnimationStateReducer'
 import { useTimeLookupFn } from 'views/common/Animation/AnimationUtilities/useTimeSyncing'
